Calculations are wrong, but approach seems okay
Kept the code of scenario 1 quite simple.  Just returning a salary map using the
data from the employee map passed as an argument.

Missing data fields in the salary map are calculated as functions that are
associated with the relevant keys.  Functions are evaluated before the salary
map is returned.

Where a function in the map uses existing values, we can use a let function to
create a local name for the result of that function.  That name can be used for
other functions in the map.